client/requirements.txt
grpcio
numpy
pandas
